fix(cms): Move seed creation into index.ts
