+++ cms/src/index.ts
@@ -1,5 +1,3 @@
-import { createUserSeed } from "./seed";
-
 export default {
 	/**
 	 * An asynchronous register function that runs before
@@ -17,17 +15,33 @@ export default {
 	 * run jobs, or perform some special logic.
 	 */
 	async bootstrap({ strapi }) {
+		const amountOfUsers = 10;
 		try {
 			const count = await strapi.entityService.count("plugin::users-permissions.user");
 			if (count === 0) {
-				createUserSeed(strapi, 10);
+				for (let i = 0; i < amountOfUsers; i++) {
+					const user = {
+						username: `dummyUser${i}`,
+						email: `dummyUser${i}@example.com`,
+						provider: "local",
+						password: "dummyPassword",
+						resetPasswordToken: null,
+						confirmationToken: null,
+						confirmed: true,
+						blocked: false,
+						role: 1,
+						old_pw_hash: null,
+						migrated_pw: false,
+						};
+						await strapi.entityService.create("plugin::users-permissions.user", {data: user});
+					}
 				console.log("SUCCESS: Created user on bootstrap");
 			} else {
 				console.log("INFO: User have already been created on bootstrap");
 			}
 		} catch (err) {
 			console.log("ERROR: Could not create user on bootstrap");
-			console.log(err.details);
+			console.log(err)
 		}
 	},
 };
